requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
cffi==1.11.2
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
Django==1.11.7
django-leaflet==0.22.0
djangorestframework==3.7.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
parsel==1.2.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.4.0
pytz==2017.3
queuelib==1.4.2
schedule==0.5.0
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.1.15
Twisted==17.9.0
w3lib==1.18.0
zope.interface==4.4.3

